requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==1.8.2
python-dotenv==1.0.1
libsql-experimental==0.0.41
requests==2.32.3